copy_data_lock & insert_nocol
Use a proper copy_data_loc, it behaves like strlcpy. It's still unclear if it's
really always a string, at any rate we guarantee null termination in case the
kernel is lying.

Handle collisions in the rank balanced tree, we just cheat time and bump it by
one until we can insert. We must be careful to not make copies of time out
before.