Link raw event in two trees
As discussed with Nick, this saves from a bad O(n) case in case we try to do
aggregation but fail since there is nothing to aggregate.

We keep one tree organized by timestamp, where RB_MIN is the oldest element,
just like before.

We keep a secondary tree organized by pid and time, so we get clusters of pid,
this is useful for assembly/aggregation where we only visit our own tree.

In the future, the pidtime will be by tidtime.