Start processing expired nodes
After the collection phase, we we start trying to assembly/aggregate, which for
now is actually just removing from the tree if they are expired.

Next step would be converting a raw_event{} to a quark_event{} and delivering it
to the user, but first we want to play with aggregation.