Lift sample_id, fixes fork hack
We want to consider fork an event of the child, it gets really confusing if
we're passing the unmodified sample_id down the stack as the user has to
remember the special case for fork.

So lift sample_type, still save the original pid of the entry in opid so that it
may be used easily for distribution in the future, or maybe hack it up as well.

Fixes a stupid comparison in the pidtime tree as well.