Stop busy waiting, do a primitive poll
wakeup_events only records PERF_SAMPLE_RECORD, we have a bunch other kind of
records, but wakeup_watermark works fine.

So use watermark for a fraction of the ring buffer size, or a fixed timeout for
now.

Since we need the data_size for calculating the watermark, better just store it,
it also simplifies the ring buffer draining code.