Introduce a quark_queue{}
quark_queue{} is the main structure the user interacts with, in the future we
want to be able to have multiple quark_queues, but that's far ahead.

Here we store all our data structures and present a uniform front to the user:

quark_{open, close, wait, get_event}... and so on

It's also a good time to do proper clean up, quark_close is the corollary to
quark_open.