Map kprobes to perf rings
Preliminary work, we can now see the events of the quark_wake_up_new_task kprobe
in the ring, we still have to handle it though.

We keep the volatile state of a kprobe in kprobe_state, and the fixed on in
kprobe, this is solely to avoid having to scrape global structures on cleanup,
making it easier to pair quark_init() and quark_close().