Aggregation in a Brave New World
This implements aggregation of fork+exec+exit:
 * Pop the oldest element in the tree (if it's old enough)
 * Look at next(by_pid_time)
   * If it can be aggregated, unlink from the tree, and link in the main event
     through agg_queue.
   * If it cannot, deliver the event.
   * If we could aggregate, also see if we can aggregate the next->next
     element(exit), do the same dance.

In order to make all of this make sense we now supress perf events by default,
it can be restored with -p for debugging.

raw_event_alloc() now must be used as we have to initialize the aggregation
queue (this could likely be a STAILQ).

raw_event_dump() was introduced and repeats a lot of the code from dump_event(),
something to be improved in the future.

While this aggregates "most test" events, we need to buffer more in order to get
them all, currently we only pop one event from each ring and try to process,
then we sleep and repeat. By the time we get to process the tree, there will be
a bunch of events sitting in the ring. This is the next step.