Use qstr to save spaces and allocations for events with pathname
A pathname is a bigboi MAXPATHLEN=4k, since a raw_event is a union of different
types, exec itself was bumping everything to 4k. Most of the time a filename is
very short, so it makes no sense to keep allocating for every small string as we
have space left in the raw_event due to being a union with some hefty members
(wake_up_new_task uses at least 2 cache lines).

qstr only allocates if size is bigger than "small", user must still call
qstr_free() on it regardless.

This is *not* a string API, I don't like custom string APIs at all, this is to
tackle this particular case, not to grow into some monstruosity with a bunch of
custom string wrappers. Use it with love, but not too much.