Introduce privilege dropping for demo purposes
This is to demonstrate we can do all we need without having any priviledges, we
open all the resources we need, file descriptors, mmap areas and whatnot, then
drop to "nobody" and chroot to "/var/empty", ideally the user would have its
home directory to be "/var/empty", but this is linux and we don't want to force
people to add a user just for a demo.

There is a joke involving a nickel and bearded dude somewhere around here.