Collect fat exit events
Tap into exit_thread (to be changed in the future as old kernels don't have the
first argument, as Nick pointed out). This gives exit code and a late snapshot
of the process as we get the full task structure, this might come in handy.

We're still missing termination time which is the next step.