Handle comm changes
While setting comm seems to be somewhere rare, there are systemd processes that
use it to signify something.

Auditbeat also uses comm as the process name entry into ECS, so it might be good
to expose to the user.

The kernel writes us no more than 16 bytes of comm, but it guarantees alignement
in the sample_id that comes just after, since the former is variable length, we
have to align it ourselves. Don't blame me.