Build dentries for cwd
This is a big change that allows us to get working_directory (cwd) up to maximum
of 7 paths.

The kernel has a limitation of 63bytes for the kprobe configuring line, the
maximum we can fit is MAX_PATH 7.

The strategy so far is to build the full path by speculating possible addresses
in the kprobe, this has limitations but it's cool and works for now.

We follow dname->d_parent until we hit mnt_root, in that case we have to start
following mnt_mountpoint.

It's likely this is not sufficient and we will have to rely on inheritance, but
it's really the best we can do with kprobes.