Get process arguments via exec_connector
This allows us to inspect the stack in order to fetch the arguments, we look at
800 bytes tops, idea from Nick.

In order to fetch the start of the stack from inside task_struct we have to dive
into an anonymous structure which our btf code wasn't prepared to handled. I've
added support for it under "(anon)", which is the same string you get from
bpftool btf dump. The issue is that we only support fetching the first anonymous
structure, if a structure has more than one and we want anything except the
first, we can't.

exec_connector will replace EXEC once we can get the filename, so hopefully
we're close to having 3 events per process, assuming we get rid of RAW_COMM
which I think we will.