Store ppid in fork
We derive ppid as the incoming pid in the originating wake_up_new_task/fork
event, therefore it is not available on exit.

The exit sample for now is a bit fatter than it should and maybe we should cut
some fields, the original idea is that it's a complete "last snapshot" of the
process metadata.

While here, make sure we set exit_code to something sensible on fork.

This is aiming at the upcoming removal of PERF_RECORD_FORK and PERF_RECORD_EXIT.