Build libquark.a
Still a bit hackish as the API is not really defined yet, but at least this is
enough so that we can really split the library from the test application, now
renamed to quark-mon, as in monitor.

Makefile might need some work when we want to have other programs for testing,
but so far it's good enough.