Introduce proper quark_events
A quark_event is the working set between the user and the library, it is a
flattened view of everything that might be aggregated on that event.

A quark_event is pretty big and will keep getting bigger, this is intended to be
allocated only on startup:

int nqevs = 64
if ((qevs = calloc(nqevs, sizeof(*qevs))) == NULL)
	err(1, "calloc");

for (;;) {
	n = quark_queue_get_events(qq, qevs, nqevs);
	if (n == -1)
		err(1, "quark_queue_get_events");
	/* Scan each event */
	for (i = 0, qev = qevs; i < n; i++, qev++)
		do_something_with_qev(qev);
}

A quark_event signals which fields have meaning trough its flags member, for
example if the event has a fork and an exit, then QUARK_EV_PROC and
QUARK_EV_EXIT will be set, meaning all proc_ and exit members are valid.

The choice for making quark_queue_get_events process a batch of events was made
to reduce the overhead of expensive FFIs, like Go, so one call = many events.
Although I'm not convinced this is necessary.