Humble beginnings of Go bindings
A QuarkEvent expresses a C.quark_event a bit differently, instead of relying on
flags to signify meaning, we just use pointers, this should be more ergonomic
for users.

We store the C representation of a number of QuarkQueue so that the hot
path doesn't allocate needlessly. The interface of QuarkQueueGetEvents is still
very similar to C, we do allocate a slice for the results though, but we can get
rid of this with channels, which might also be more ergonomic.

Blocking from cgo is evil and the runtime scheduler doesn't cope with it very
well. So we avoid it by doing the polling on the Go side.

Handling errno in Go is a bit odd, we should only trust errno if we get an error
in the primary value. There is also a chance that there are paths in quark that
don't properly set errno, while in the C side this is not a big deal as we just
get a bad error message, in the Go side this would signify err == nil, which is
pretty evil. So wrap all remaining cases in ErrUndefined. There are likely
better ways to do this but for now it will do.

In my naive tests (while true; do whoami>/dev/null;done), Go performs as well as
the pure C, but it's also not doing the complete printing yet.