Implement an event cache
The cache allows us to enrich each small event with everything we know about the
process, say for example you got a fork+exec, fork then spills out an event with
all events it knows from fork+exec. Sometime later that same process does
something small, like setproctitle(), we can then send out the COMM event with
all previously known information from fork+exec.

Also included is a user function for querying the cache of a process, this
should be useful for users who need to enrich some other thing with process
information.

If the process is exiting, we give it a grace time before purging it, this
allows the user to still query it for some time, which would very likely be the
case on short lived processes.

Still missing is purging/replacing the cache in case of a PID collision, for
this, event_cache_insert() needs to be passed a flag on when to replace things.

We always fill in the cache and then send out to the user, the cache can be
disabled by passing QQ_NO_CACHE to quark_queue_open().