Fix Go bindings due to bad cgo check
When you make a CGO call on a pointer that the runtime knows the structure, CGO
traverses all nested members looking for pointers that point back to the Go
heap. The problem is when you have members that reference themselves, CGO
wrongly thinks you're pointing to some GO allocation, while in fact you're just
pointing to "yourself". This would obviously be an issue if the C code keeps
pointers to the Go Heap, which it doesn't in quark.

This is due to us now having an empty TAILQ_HEAD() inside struct quark_event{},
after TAILQ_INIT(), next points to the address of head:
	TAILQ_FIRST((head)) = NULL;					\
--->	(head)->tqh_last = &TAILQ_FIRST((head));			\
	QMD_TRACE_HEAD(head);						\
	} while (0)

Imagine the following:

func crashQuark() {
	var qq C.struct_foobar;

	C.quark_open(&qq, 0) <-- this is ok, qq.event_gc.tqh_last is NULL
        ...
	C.quark_close(&qq) <--- this will panic
	qq.event_gc.tqh_last points to &qq.event_gc.tqh_head
	which is pretty much a Go address since qq itself is in the Go heap, but
	there is nothing wrong with it, cgo is just really aggressive. It recursively
	dives into every member of the structure checking for pointer types that point
	to the Go heap, even if that address is only pointed to "itself".
}

The solution is to use opaque pointers. Again there is nothing wrong, the
structure is just pointing to itself! It's not pointing to any outside value
that resides in the Go Heap.

The solution is to use opaque pointers, like most people do.
An alternative would be to pin qq.event_gc.tqh_last before every call, but
that's terrible we would also have to make sure to only do it when it's not
self.