Implement snapshots (/proc scraping)
This change primes the event cache with process information retrieved via
scraping /proc.

Events will show up as QUARK_EV_SNAPSHOT to the user and are retrieved via
quark_get_events() in the same way as "edge" events. In order to do this,
quark_get_events() now needs to be smart enough to know where it is on the
snapshot, as a snapshot ends up being transferred in multiple quark_get_events()
calls.

It might make sense to not take a snapshot on quark_queue_open(), and instead
give the user a button like quark_trigger_snapshot(), this could be useful on
cases where the user opens the queue but only listen to events much later.

We open up the queue rings before doing the snapshot, so we can recover from a
possible missed event. In the worst case, we will see a process showing up
twice.

Some helper functions were also adjusted to be a bit safer.