Revamp command line args. Issue #8
Change how we represent cmdline in quark_event, now we just pass the flattened
value up, and it's the user responsability to put it into something meaningful.
args_to_spaces and args_make are provided to easy the burden. While here rewrite
args_to_spaces into something simpler.

Add two forms of printing to quark_event_dump for cmdline, leave one out for now
until we have a proper argument to toggle.

In Go we do the conversion manually as to not call into cgo.