Implement inheritance. Issue #10
Inherit COMM, FILENAME and CMDLINE from parent, as we were only getting those
values from exec before, forks without execs were a bit lonely. We could inherit
all fields but as we know we are just about to overwrite them, it's better to
single out the ones we really need.