EBPF Support. Issue #13
This adds EBPF as a backend, based on the EBPF programs from elastic/ebpf.

While the diff appears huge, this is mostly due to splitting quark.c into
bpf_queue.c and kprobe_queue.c. The upper layer, aggregation and whatnot remains
unchanged.

Backend expression is done through queue_ops, so one for EBPF and one for
KPROBE.

We load the EBPF programs ourselves through bpf_queue.c, and it uses the libbpf
skeleton thingy, that means we also need to generate the skeleton, which implies
having bpftool(8) in path.

Initialization was also changed and quark_init(3) and quark_close(3) have been
removed, this has been in my mind for some time and with EBPF it became clear it
was the right thing to do. Code from quark_init() and quark_close() has been
merged into quark_queue_open() and quark_queue_close() which now grows in
complexity as it needs to keep state for multiple queues, but at the same time
makes it atomic to initialize/fail on a specific backend.

Unless the user requested a specific backend in quark_queue_open(), it will
attempt to use EBPF and then fallback to KPROBE.

RAW_EXEC has been extended with a RAW_EXEC_F_EXT that is used by EBPF, while in
kprobes we can't get all we need from one RAW_EXEC, in EBPF we do, so use the
same structure and signal that the extra fields are good.

raw_event_to_quark_event() has been improved to save some code, as now we can
get the same field from different places, it makes sense to unify the code that
handles it.

Skeleton loading makes valgrind a bit dirty since it does a syscall without
clearing up the stack, this is a diff for upstream as it's a oneliner.

We had to settle in CLOCK_MONOTONIC as we don't have access to
CLOCK_MONOTONIC_RAW from EBPF, this was coming as older kernels wouldn't support
that in kprobes as well.

STILL MISSING:
The extra capabilities from process are not available and I haven't added them
to elastic/ebpf as it's likely uneeded, so now it's just blank on EBPF, the
important ones are there though.

It's still missing a counter of missed events for EBPF that must be added in
elastic/ebpf as we have to do the counting ourselves, this is probably the next
thing to tackle.