Don't restart on EINTR in Go Block
Let the code try to get some data, in case we ever change the timeout to be
bigger, it might be that we always get an interrupt before the timeout expiring,
which would case us to never check for events.

The Go scheduler uses signals to force fiber switching, and we can't control
that.