Import elftoolchain/libelf verbatim. Issue #7
This is a partial import of elftoolchain/libelf with only the files we are
really going to use.

This commit has no modification to any of the files, except for a new file
`commit` added so that we can track our starting point, it tracks the commit of
upstream elftoolchain.

A non-complete list of what has NOT been imported:
	- addr2line/
	- ar/
	- as/
	- brandelf/
	- cxxfilt/
	- documentation/
	- elfcopy/
	- elfdump/
	- findtextrel/
	- GNUmakefile
	- INSTALL
	- isa/
	- ld/
	- libdwarf/
	- libelftc/
	- libpe/
	- Makefile
	- mk/
	- nm/
	- readelf/
	- README.rst
	- RELEASE-NOTES
	- size/
	- strings/
	- test/
	- tools/
	- libelf/{Makefile, *.3, *.cat3}
	- common/Makefile
	- common/sys/Makefile

The next commit has all the changes and additions we did to make this work
nicely with quark and a recent libbpf.