Hook libz into Quark. Issue #7
This hooks up zlib into the build system, no changes to source code, just tweaks
to the makefile to make it "look" like the rest of the build.

make V=1 should restore all the original output, like the others.

Initially I've tried this as a submodule, but zlib is a bit weird as it includes
some autogenerated files in the commit log, then it gets regenerated at build
and shows up as modified.