Get session ID. Issue #26
This adds session id on fork for kprobes and ebpf.
We still need to track changes through a process calling setsid(2), and then
do the same for cred changes.

Parser for the kprobe DSL had to be extended to accept recursion after an
arithmetic expression.
It's a bit ugly as we're stretching the DSL, life is hard.