Add attr.{cache_grace_time,hold_time,QQ_NO_SNAPSHOT}
cache_grace_time is how long we keep an event in cache after the process exited.
hold_time is how long we buffer an event for aggregation and ordering.
QQ_NO_SNAPSHOT supresses the initial snapshot via /proc scraping.

The user facing hold_time is in MS, and the internal facing hold_time is in NS,
as long as people are following the docs it should be safe, but it is a bit
confusing.

Maybe we should cut all the NS inside and lose precision, I had a good reason in
the beginning for keeping it all in ns, but I can't remember.

QQ_NO_SNAPSHOT is useful for debugging and seeing what we're inheriting and what
we're learning from individual processes, I doubt anyone would ever want to use
this in production, but it's a simple feature that aids debugging.