Add TTY major, minor. Issue #32
This should be enough for the existing state of session processor provider in
beats.

It's worth noting that minor is actually larger than 16bits these days, proc(5)
states:

              (7) tty_ %d
                     The controlling terminal of the process.  (The minor
                     device number is contained in the combination of bits 31
                     to 20 and 7 to 0; the major device number is in bits 15
                     to 8.)

That's 20 bits for minor, the kernel aggrees:

static __always_inline u32 new_encode_dev(dev_t dev)
{
	unsigned major = MAJOR(dev);
	unsigned minor = MINOR(dev);
	return (minor & 0xff) | (major << 8) | ((minor & ~0xff) << 12);
}

With kprobes, for kernel threads, task->signal->tty is NULL, which results in us
reading garbage for tty major and write, that's why we need tty_addr. In case
it's NULL, we don't read tty_major and tty_minor.

EBPF is missing tty major/minor for fork and must be upstreamed.