Compile btfhub-archive offsets. Issue #22
This diff adds the infrastructure necessary for handling btf files from the
btfhub-archive: https://github.com/aquasecurity/btfhub-archive

If we're running on a kernel that has no BTF support, we will try to use a
matching btf from the archive.

This covers ancient kernels (since the 3.x era), so a lot has changed, follows a
summary of how we try to handle them:

Some members simply changed names within the same structure, this is the case of:

struct btf_alternative {
	const char *new;
	const char *old;
} btf_alternatives[] = {
	{ "task_struct.start_boottime",		"task_struct.real_start_time"	},
	{ "mm_struct.(anon).start_stack",	"mm_struct.start_stack"		},
	{ NULL,					NULL				},
};

In order to cope, we just try the alternative name and see if the offset
resolves, if it does, win.

The problem arises when the actual linkage changes, for example task_struct.pids
moved to signal_struct.pids in newer kernels, so we can't just look at an
alternative name. To handle this we kludge the actual kprobe_dsl at runtime, for
example pgid suffered from the pids change, it's now defined as:

{ "pgid",		"di", "u32",	"KLUDGE - see kprobe_kludge_arg()" }

The definition for new kernels remain at:
struct kprobe_arg ka_task_new_pgid = {
	"pgid", "di", "u32", "task_struct.group_leader task_struct.signal (signal_struct.pids+16) (pid.numbers+0).upid.nr"
};

and old kernels at:
struct kprobe_arg ka_task_old_pgid = {
	"pgid", "di", "u32", "task_struct.group_leader (task_struct.pids+8) (pid.numbers+0).upid.nr"
};

When we're processing the dsl, we have a function that attempts to detect which
offsets are present and then decides on using one or the other, this is
intentionally very manual as we might have to replace "everything". The
function is kprobe_kludge_arg().

Now quark_btf{} ends up trying to resolve offsets that will always fail (if
task_struct.pids is present, signal_struct.pids is not), so I had to relax the
parsing, which before required every offset to succeed, now one of these 2 might
fail and we ignore, if anything else fails we still bail.

btfhub.c is a big fat table with a quark_btf{} for each btf file we parse, for
amd64 alone it's ~3k files and btfhub.c ends up with 120k lines, the object file
is 4.7M when not stripped, half of that stripped. We don't strip for now.

The generation has a proper make target:

make btfhub BTFHUB_ARCHIVE_PATH=/d/btfhub-archive

To generate the actual C code in btfhub.c a new option has been added to
quark-btf(8), with -g we can trigger the whole dance.

btfhub-archive *does* change, in span of 4 months they added ~50 new entries, so
there will be future use.

This commit really only generates the offsets and makes sure we can resolve all
that quark needs, it doesn't use them yet, that's step 2.