Use local libbpf includes. Issue #46
This prevents quark from using the system libbpf headers, it should use the ones
from the included libbpf src.

Sadly libbpf doesn't have an internal layout that matches the headers, as in: it
doesn't store the headers in the same directory layout they are installed, example:

./libbpf/src/libbpf.h -> <include-prefix>/bpf/libbpf.h

We need that "bpf" in there so we sadly have to copy them to a "build include"
directory, this way code can include them as they would if they were using
system headers:

On the positive side it kills the horrible sed(1) call.