API for dumping the internal cache as a dot file. Issue #39 (#57)
API for dumping the internal cache as a dot file. Issue #39

This also cleans up a bit of the older dot files.

This dumps every process in the cache and joins siblings by color, which makes
it a bit easier to understand what is going on. This is mostly a debugging tool,
but decent enough to be exported. manpage for it pending.

Co-authored-by: Nicholas Berlin <56366649+nicholasberlin@users.noreply.github.com>