Docker builds and initial arm64 cross compilation
This makes quark "easily" buildable from a docker container, it also cross
builds and links arm64, but it shouldn't really work at the time.

To get the ball rolling you do:

$ make docker

It should built the image and then build quark.
At the time, containerized builds always imply a `cleanall`, this is temporary
until I find a solution I don't hate.

In order to debug the container you can do the following, which puts you in an
interactive shell inside the container.

$ make docker-shell