Refactor API and quark_event{} 1/3. The meaty chunk
Previously, quark always copied stuff out from internal structures to a user
storage, this works fine and it's easy to reason about it when the
datastructures are mostly fixed sized.

This refactor moves most of what was previous in quark_event{} to a new
structure called quark_process{}.

There's two major reasons for this change.

1 - In the near future we will want to add file descriptor tracking to quark,
which is inheritently not fixed size, and adding an extra API call to handle
file descriptors doesn't look nice.

2 - Stuff just looked weird. For example quark_event_lookup(3) would lookup a
quark_event{}, which is in fact just a copy of the internal process cache, the
reasoning it's that it's "all the events that happened to it". So previously an
event and a process were essentially the same.

In this change, the `events` field of old quark_event{} is kept and it expresses
the same thing as before, but all other fields were moved to quark_process{},
and quark_event{} now gets a pointer to the internal process structure from
quark. Meaning now we don't copy stuff out anymore, we return pointers to
internal data, yadayadayada more dangerous, but we're all adults here, and also
zero copy. The main reason is it's more ergonomic, just look at how
quark_process_iter{} is much nicer now.

Reviewing this is painful, most of the changes is renaming the variable `qev` to
`qp`, as in `quark_process`. The real changes are here and should be the focus of
the review:
 - raw_event_process(), previously raw_event_to_quark_event(). We don't copy the
whole event anymore, we just point the passed `dst` to a an internal
quark_process{} and fill `events` with the corresponding flags.
 - quark_event_iter{}, doesn't read terrible anymore and doesn't do one lookup
for each entry, so now O(n), previous O(n) * O(log n).
 - QQ_NO_CACHE didn't survive this change, he is going back to his family in the
south. We have no storage on where to put the process_cache{} needed for
QQ_NO_CACHE, so I killed it as I don't want memory management between quark and
the Api.
 - event_copy_out() doesn't do anything anymore and was commented out, it might
come back as an exported function for copying an event out, so keep it for awhile.