Prepare documentation for github pages. 1/2
Github pages can be derived from a source directory on a source branch. The
source directory is actually either `root` or `docs`, so we're left with docs.

This means now we will generate the html files in docs/ and commit them, as
before we didn't commit. Also we properly track the dependency of the html to
the manpage now, cause it's cool.

This commit removes all doc targets in favor of one true `docs`, it also does
the linting in the same target now.