Support older perf rings, needed for centos7
This is one step into making centos7 kprobes work, still missing the btfhub PR,
but I've tested and it works, trust me, I'm an engineer.

centos7 perf_rings don't support clock_id and exec_comm, it might be that we
need to add something back when exec_comm is removed, but for now let us get the
probes to at least install.

The offsets for `task_struct.pids` were wrong, I got mistaken in my first
attempt as in newer kernels that structure is an array of pointers, but there
it's a struct of pid_link:

https://elixir.bootlin.com/linux/v3.10/source/include/linux/pid.h#L69

and sizeof(struct pid_link) is 24: 16 for hlist_node, 8 for the pid pointer.
This took me awhile to see because they declare the two members of hlist_node in
the same line, disappointing.