Zap Go, moved to https://github.com/elastic/go-quark
We moved the Go files to another repository as we have to ship the archives in
order to ease go distribution.

If we kept the files here, we would have to rebuild and commit them at every
step of development which would bloat the repository size considerably.

In go-quark we will only update quark itself when we're about to release.