Add three more tests to quark-test(8)
This adds the following tests:
  * t_cache_grace: ensure processes are gced after grace time.
  * t_min_agg: ensure minimal aggregation works.
  * t_stats: ensure stats (counters) are working.

Tests are ordered by increasing order of complexity so we can see what fails
earlier. We could assert the whole event like we do in t_fork_exec_exit(), and
we might do it in the future after factoring it out to a separate function.