Add test targets and scripts
This adds targets `test` `test-kernel` and `test-all` along with the scripts
supporting them `krun.sh` and `ktest-all.sh`.

`krun.sh` can be used to run a quark utility in a specified kernel:

$ ./krun.sh initramfs.gz kernel-images/amd64/linux-6.11.4-101.fc39.x86_64 quark-test -v

`ktest-all.sh` is just a wrapper for running all kernels in kernel-images.

Keeping the idea of not running scripts directly, these will mostly be used via
`test-kernel`.