Plug a missing build dependency
libbpf depends on elfdefinitions.h which is generated when we build the
elftoolchain, ideally we would have a target for the headers only, but that's a
lot of work for zero gain, so just make libbpf depend on building the whole
elftoolchain.

Makes it possible to do a clean checkout and run something like
`make quark-test` without doing `make all first`.