Add namespace inode numbers to process events
This adds the inode numbers of 4 namespaces to both kprobe and ebpf backends:
 - uts
 - ipc
 - mnt
 - net

The interesting ones we are missing are pid and cgroups, the former is a bit
complex to fetch from kprobes and the later is not present in all kernels. Since
we don't have use for them at the moment, don't bother trying to include them.

The only real needed one at the moment is mnt so that auditbeat can check if it
is in the same mnt namespace as the process.

net offset location changed twice, so we need two alternatives instead of one,
but the existing infrastructure could not cope with two alternatives, so kill
btf_alternative_name() and loop over all possible alternatives to support it.

Note that if we fail to parse the sproc namespace entry, maybe cause the process
just vanished, we store zero which is an invalid inode number, the race recovers
as we only parse /sproc with the rings opened, meaning we will likely see the
exit event, which has the correct inodes.