Fix a race in ktest-all and improve output
Turns out I broke it on last commit when there is more than one kernel to test,
both processes were attempting to read from stdin, so revert that and fix it
differently.

While here, improve some idioms and collect errored out commands to be printed
in the end, should make people less angry in the future when stuff just "fails".