Allow quark to be linked against system libraries
This adds a SYSLIB=y option to quark to use the installed system libraries instead of building its own
libbpf,zlib,elftoolchain.

It's too much hassle to compile stuff statically in the SYSLIB case so disable
the *-static and initramfs targets.

While here, simplify all and allow it to rely solely on each of the final
targets dependency, which uncovered that zconf.h was not an explicit dependency
of libbpf, so fix that as well.