Replace quark_queue_get_events(3) with get_event(). Fixes #107
quark_queue_get_events(3) was initially designed when I wanted to copy
everything out of quark, which was changed in
9d38ecf767a02a080ae8ed7bccf66ecf2281b3ee to return pointers to internal state.

The problem is that since that change we were getting "future" state on events
of the same pid, on a quark_queue_get_events() == 2, we could have:

[0] pid 7 exec foo
[1] pid 7 exec bar

event [0] will show up with filename "bar" instead of "foo", since they update
the internal cache and return the pointer to it.

This diff kills quark_queue_get_events(3) and replaces with a single
quark_queue_get_event().

With this, quark is able to produce 3.4 million events per second on my laptop,
so I think we're fine. If I find that this affects Go too much, which my initial
tests show it doesn't, I can bring quark_queue_get_events(3) back in a way that
copies out everything, like the first version.